Change count to uint8_t

+++ i2c/sample-implementations/GPIO_bit_banging/sensirion_i2c_hal.c
@@ -88,10 +88,10 @@ void sensirion_i2c_hal_free(void) {
  * @param count   number of bytes to read from I2C and store in the buffer
  * @returns 0 on success, error code otherwise
  */
-int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint16_t count) {
+int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint8_t count) {
     int8_t ret;
     uint8_t send_ack;
-    uint16_t i;
+    uint8_t i;
 
     ret = sensirion_i2c_gpio_start();
     if (ret != NO_ERROR)
@@ -123,9 +123,9 @@ int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint16_t count) {
  * @returns 0 on success, error code otherwise
  */
 int8_t sensirion_i2c_hal_write(uint8_t address, const uint8_t* data,
-                               uint16_t count) {
+                               uint8_t count) {
     int8_t ret;
-    uint16_t i;
+    uint8_t i;
 
     ret = sensirion_i2c_gpio_start();
     if (ret != NO_ERROR)
